pyproject.toml
[tool.poetry]
name = "sun-kist"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "sun"}]

[tool.poetry.dependencies]
python = "^3.11"
llmlingua = "^0.2.2"
openai = "^1.86.0"
httpx = "0.27.2"
langchain = "^0.3.26"
langchain-openai = "^0.3.28"
langchain-chroma = "^0.2.4"
langchain-text-splitters = "^0.3.8"
langchain-community = "^0.3.27"
tavily-python = "^0.7.10"
beautifulsoup4 = "^4.12.0"
langgraph = "^0.5.4"
playwright = "^1.54.0"
numpy = ">=2.1.0"
pillow = "^11.3.0"
streamlit = "^1.47.0"
python-dotenv = "^1.1.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
